+++ neat/locals/vm_selection/algorithms.py
@@ -33,7 +33,7 @@ def random_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
@@ -52,7 +52,7 @@ def minimum_utilization_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
@@ -72,7 +72,7 @@ def minimum_migration_time_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
